refactor(proactive): 迁出主动决策收尾
